% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cogarch_sim.R
\name{cogarch_sim}
\alias{cogarch_sim}
\title{Simulation of COGARH(1,1) process}
\usage{
cogarch_sim(t = 0:10, beta = 1, eta = 0.05, phi = 0.03, Lp = "cp",
  rate = 1, distribution = "normal", mean = 0, var = 1,
  vg_param = list(sigma = 1, nu = 0.5, theta = 1), gs = 0.01)
}
\arguments{
\item{t}{fixed time grid}

\item{beta}{parameter of the cogarch process}

\item{eta}{parameter of the cogarch process}

\item{phi}{parameter of the cogarch process}

\item{Lp}{driving Levy process: either "cp"=compound Poisson or ""vg"= Variance-Gamma}

\item{rate}{intensity of the compound Poisson Levy process}

\item{distribution}{jump size distribution, e.g "normal"}

\item{mean}{mean of specified jump distribution}

\item{var}{variance of specified jump distribution}

\item{vg_param}{list of parameters for Variance-Gamma process
(for E(L_1)=0 and E(L_1^2)=1 choose theta=0 and sigma=1)}

\item{gs}{time grid size}
}
\value{
list containing the COGARCH(1,1), volatility and driving Levy process. As well as
  the jumps of the Levy process and the random jumptimes
}
\description{
cogarch_sim simulates a COGARCH(1,1) process driven by a compound Poisson or a Variance-Gamma
Levy process
}
\details{
The function use the Euler-Maruyama scheme in order to get a numerical
solution for a SDE.
See the Book: Peter E. Kloeden: "Numerical solution of stochastic differential equations"
}

