% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driv_levy_processes.R
\name{comp_pois_proc}
\alias{comp_pois_proc}
\title{Simulation of a compound Poisson process}
\usage{
comp_pois_proc(start = 0, end = 10, lambda = 1, distribution,
  dparams = list(), obs_times = start:end)
}
\arguments{
\item{start}{start point of the time interval [start,end]}

\item{end}{end point of the time interval [start, end]}

\item{lambda}{intensity of compound Poisson process}

\item{distribution}{expression defining the distribution of the jump sizes
in the form r'dist'}

\item{dparams}{further parameters of the jump size distribution}

\item{obs_times}{observation times of the process in the intervall [start, end]}
}
\value{
output list containing the jump times, the jump sizes, the observation times
  and the compound Poisson process
}
\description{
comp_pois_proc simulates the trajectory of a compound Poisson process
}
\details{
The jumps of the compound Poisson process arrive accordingly to a poisson process.
The sizes of jumps are random, with a specified probability distribution.
}
\examples{
comp_pois_proc(distribution = rnorm, dparams = list(mean = 0, sd = 2))
comp_pois_proc(end = 100, obs_times = seq(1/10, to = 100, by = 1/10),
                 distribution = rt, dparams = list(df = 2))
}

