% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driv_levy_processes.R
\name{nig_proc}
\alias{nig_proc}
\title{Simulation of a NIG process}
\usage{
nig_proc(sigma, kappa, theta, obs_times)
}
\arguments{
\item{sigma}{standard deviation of the Brownian motion}

\item{kappa}{intensity of the subordinator}

\item{theta}{drift parameter}

\item{obs_times}{observation times of the process}
}
\value{
output list containing the jump times, the jump sizes, the observation times
  and the NIG process
}
\description{
nig_proc simulates the trajectory of a normal inverse Gaussian process
with parameters sigma>0, kappa>0 and theta
}
\details{
The process is defined ...
}

