% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driv_levy_processes.R
\name{var_gam_proc}
\alias{var_gam_proc}
\title{Simulation of a Variance Gamma process}
\usage{
var_gam_proc(sigma, nu, theta, obs_times)
}
\arguments{
\item{sigma}{standard deviation of the Brownian motion}

\item{nu}{scale parameter of the Gamma distribution defining the increments
of the Gamma process}

\item{theta}{drift parameter}

\item{obs_times}{observation times of the process}
}
\value{
output list containing the jump times, the jump sizes, the observation times
  and the Variance Gamma process
}
\description{
var_gam_proc simulates the trajectory of a variance gamma process
with parameters sigma>0, nu>0 and theta
}
\details{
The process is defined through a time change TG, which is done wrt a gamma process.
W_TG is then a time changed standard Brownian Motion and the variance gamma process is
defined through
     V = theta*TG + sigma * W_TG
}

